Reworking entity system part 2
Camera is now a core "thing" and it's no longer an entity (it was never a component, just a special entity) and can be accesseble from the scene.
The editor Grid is now an Widget and it is special in the sense of it has the custom vulkan drawcalls and the imgui draw call. Customized ui elements are now possible.